perft works with uint64

+++ includes/move_maker.hpp
@@ -10,7 +10,7 @@ void MakeMove(Move mv);
 /// the position and the stateinfo undoing the move
 void UndoMove(Move mv);
 
-// count legal moves
-int StupidPerftCount(MoveList mvList, uint8_t depth,uint8_t MaxDepth, bool verbose = false, int count = 0);
-// count legal moves
-int StupidPerftCount(std::string FEN, uint8_t MAxDepth, bool verbose = false);
+/// count legal moves only at depth, non cumulative. Mimicks stockfish
+uint64_t StupidPerftCount(MoveList mvList, uint8_t depth,uint8_t MaxDepth, bool verbose = false, uint64_t count = 0);
+/// count legal moves only at depth, non cumulative. Mimicks stockfish
+uint64_t StupidPerftCount(std::string FEN, uint8_t MAxDepth, bool verbose = false);
